Reorganize chunk pool code structure & fmt codes
